Relax triangulation constraints
This changes the behaviour of the triangulation for estimating the
lat&lng of images lacking geolocation data so that we only require one
of the neighbours (prev or next image in temporal order) to provide an
estimated location. In case only one of them is present, assuming the
previous/next image was also shot relatively close (time-wise) to the
current image's time, we will reuse the same coordinates. The behaviour
is left unchanged in case both the previous and the next image are
present and include correct geolocation information.